fix cmpr
