--- BitFaster.Caching.UnitTests/Lfu/CMSketchAvx2Tests.cs
@@ -1,7 +1,0 @@
-﻿
-namespace BitFaster.Caching.UnitTests.Lfu
-{
-    public class CMSketchAvx2Tests : CmSketchTestBase<DetectAvx2>
-    {
-    }
-}
