--- BitFaster.Caching.UnitTests/Lfu/CmSketchTestBase.cs
@@ -1,97 +1,0 @@
-﻿
-using System.Collections.Generic;
-using System.Runtime.Intrinsics.X86;
-using BitFaster.Caching.Lfu;
-using FluentAssertions;
-using Xunit;
-
-namespace BitFaster.Caching.UnitTests.Lfu
-{
-    public abstract class CmSketchTestBase<TAVX> where TAVX : struct, IAvx2Toggle
-    {
-        private CmSketch<int, TAVX> sketch = new CmSketch<int, TAVX>(512, EqualityComparer<int>.Default);
-
-        public CmSketchTestBase()
-        {
-            SkipAvxIfNotSupported();
-        }
-
-        [SkippableFact]
-        public void WhenCapacityIsZeroDefaultsSelected()
-        {
-            sketch = new CmSketch<int, TAVX>(0, EqualityComparer<int>.Default);
-
-            sketch.ResetSampleSize.Should().Be(10);
-        }
-
-        [SkippableFact]
-        public void WhenIncrementedOnceCountIsOne()
-        {
-            sketch.Increment(1);
-            sketch.EstimateFrequency(1).Should().Be(1);
-        }
-
-        [SkippableFact]
-        public void WhenIncrementedMoreThanMaxCountIsMaximum()
-        {
-            for (int i = 0; i < 20; i++)
-            {
-                sketch.Increment(1);
-            }
-
-            sketch.EstimateFrequency(1).Should().Be(15);
-        }
-
-        [SkippableFact]
-        public void WhenTwoItemsIncrementedCountIsIndependent()
-        {
-            sketch.Increment(1);
-            sketch.Increment(1);
-            sketch.Increment(2);
-
-            sketch.EstimateFrequency(1).Should().Be(2);
-            sketch.EstimateFrequency(2).Should().Be(1);
-        }
-
-        [SkippableFact]
-        public void WhenSampleSizeExceededCountIsReset()
-        {
-            bool reset = false;
-
-            sketch = new CmSketch<int, TAVX>(64, EqualityComparer<int>.Default);
-
-            for (int i = 1; i < 20 * 64; i++)
-            {
-                sketch.Increment(i);
-
-                if (sketch.Size != i)
-                {
-                    reset = true;
-                    break;
-                }
-            }
-
-            reset.Should().BeTrue();
-            sketch.Size.Should().BeLessThan(10 * 64);
-        }
-
-        [SkippableFact]
-        public void WhenClearedCountIsReset()
-        {
-            sketch.Increment(1);
-            sketch.Increment(1);
-            sketch.Increment(2);
-
-            sketch.Clear();
-
-            sketch.EstimateFrequency(1).Should().Be(0);
-            sketch.EstimateFrequency(2).Should().Be(0);
-        }
-
-        private static void SkipAvxIfNotSupported()
-        {
-            // when we are trying to test Avx2, skip the test if it's not supported
-            Skip.If(typeof(TAVX) == typeof(DetectAvx2) && !Avx2.IsSupported);
-        }
-    }
-}
